frontend/requirements.txt
Flask==1.0.1
Flask-WTF==0.14.2
requests==2.20.0
Flask-SQLAlchemy==2.3.2
